edit-vehicle-fix
