--- app/views/users/update.html.erb
@@ -1,2 +1,0 @@
-<h1>Users#update</h1>
-<p>Find me in app/views/users/update.html.erb</p>
