--- test/controllers/users_controller_test.rb
@@ -1,13 +1,0 @@
-require "test_helper"
-
-class UsersControllerTest < ActionDispatch::IntegrationTest
-  test "should get edit" do
-    get users_edit_url
-    assert_response :success
-  end
-
-  test "should get update" do
-    get users_update_url
-    assert_response :success
-  end
-end
